Make cloning possible
Cloning of a board enables us two important things: with this we can detect an (fairly typical) endless loop in the algorithm, because we can calculate a unique hash of a board, and maintain a list of board layouts that were already checked.
The second very important improvement is that with cloning it will be possible to design multi-threaded routing algorithms and improve the efficiency and speed significantly.